+++ examples/parl_io.rs
@@ -231,7 +231,7 @@ async fn hub75_task(
         latch: peripherals.latch,
     };
 
-    let mut hub75 = Hub75Type::new_async(peripherals.parl_io, pins, channel, tx_descriptors);
+    let mut hub75 = Hub75Type::new_async(peripherals.parl_io, pins, channel, tx_descriptors, 15.MHz());
 
     let mut count = 0u32;
     let mut start = Instant::now();